Split PCAQuasar into superclass, PCAQuasarScoringBench & PCAQuasarTrainingBench
